requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
google-generativeai>=0.8.0
Pillow==9.5.0
requests>=2.31.0
moviepy==1.0.3
pydub==0.25.1
openai-whisper
pysrt==1.1.2
pymongo==4.16.0
flask-login==0.6.3
flask-bcrypt==1.0.1
dnspython==2.8.0
numba>=0.58.0
llvmlite>=0.40.0
gunicorn==21.2.0
